Rename file demo-section.vue to DemoSection.vue
